requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
jinja2==3.1.4
python-multipart==0.0.9
playwright==1.47.0
beautifulsoup4==4.12.3
lxml==5.3.0
sqlmodel==0.0.22
pydantic-settings==2.4.0
httpx==0.27.2

PyYAML==6.0.2

PyYAML==6.0.2
